feat: Ensure codegen files are type checked.
We want users to be able to have skipLibCheck: true in their
tsconfig. However, we do need to ensure that our codegen files are
not skipped in such cases. So we no longer use the .d.ts extension.

"Generate Godot d.ts" is now "Generate types" and in addition to
generating all Godot and project types (which it already did), the
autogen directory will now wiped of all files/directories before
commencing generation. This ensures old generated files no longer
pollute the project.