feat: GArray tuple support.
Much like JS Array tuples, there are a plethora of APIs that are
able to mutate the GArray tuple to violate the tuple typing. The
intent is that only basic set/get functions (which are type safe
for tuples) ought to be used.

For now I've opted not to add support for tuples to GArray's
proxy() types. Simply because the types involved are already very
complicated/slow. Perhaps this will be revisited in future.