fix: Don't crash on GodotJS native class mismatch.
Instead we log an error and prevent the script from being
instantiated. We probably shouldn't outright crash due to user
error in general. However, beyond that, this is actually an
expected state when using the editor. If you wanted to change node
type, then without this fix you have a chicken and egg problem
between the script and the scene.