fix: JavaScriptTimerAction invocation crash.
Previously, a crash would occur on the following line:

```cpp
const v8::Local<v8::Context> context = func->GetCreationContextChecked();
```

This is because SArray does not copy/move upon resizing, instead
the raw memory is copied. However, v8::Global (where func came from)
is not POD. You fundamentally cannot move them around in memory
because their design fundamentally requires that V8 keep a running
record of all references for when the slot is moved (with
generational GC, this sort of thing happens frequently).

Solution is to allocate our v8::Global on the heap and store a
pointer to it. The pointer can safely be moved around in memory
without issue.