fix: Actually fix JavaScriptTimerAction invocation crash
Previous fix was a red herring. Whilst technically there was an
issue there, it turned out not the be the cause of the crash that
I've repeatedly ran into.

The real issue was that SArray isn't really a safe data structure,
by design. Smaller revision ranges exacerbate the data structure's
limitations. For some use cases a small revision range will be
fine, but timers are likely a prime example of an API where a small
revision range is NOT fine.

This is because the way timers are used in JavaScript is often as
timeouts, which is why the API is called setTimeout(). Often the
user schedules a timeout and it's never expected to fire. Instead
clearTimeout will be called when some other operation returns
before the timeout is reached. This means timers are frequently
being inserted and deleted. SArray by design will reuse indexes,
when this occurs the revision is incremented. However, the Index32
data structure only permits 64 revisions. That number is likely to
(and in my case definitely does) wrap around frequently.

That on its lonesome isn't an issue. What's problematic is that
the timer manager uses an efficient wheel data structure built
atop SArray. Scheduled timers are added to one of many wheels and
into the SArray index of all scheduled timers. When a timer is
cleared, the wheels are not touched at all, only the used timer
SArray is modified. The idea being that later on when it's time
to execute the relevant timers stored in (part of) a wheel, we
just check to see if index in the wheel resolves to a timer in
our used/scheduled timer SArray. We a small revision range, we
frequently end up reusing both and index AND matching revision.

This leads to both early firing of timers AND firing of timers
that have since been destructed i.e. cause the crash.

This commit swap from using 32-bit index+revisions to 52-bit
index+revisions. 20 bits of which are reserved for the revision.
So now we can have over 1 million revisions before wrapping
around.

It's now substantially unlikely for index re-use to happen.
HOWEVER, in theory, it COULD still occur. Suppose that in this
extremely unlikely circumstance our timer were to fire too
early. That's not ideal, but is probably a risk we can live with.
What we can't accept is the risk of an outright crash. So
we've also implemented protection in JavaScriptTimerAction
itself. Basically its destructor now effectively marks itself as
being destroyed. The execution logic detects this and bails out,
evading the crash.